Fixed delete stuff a bit

+++ upload_coordinator/src/endpoints/support_delete_all_cases_files_case_uuid.ts
@@ -1,31 +1,53 @@
 import { Request, Response } from "express";
 import { execute_sql } from "../sql";
-import {
-    SignedMessage,
-    signMessage,
-} from "../../../backend/src/endpoints/uploads/signing";
+import { createVerify, createSign } from "crypto";
 
 const UPLOAD_SERVICE = "itassets-coordinator-service:3000";
 
 export async function delete_support_delete_all_cases_files_case_uuid(
     req: Request,
-    res: Response
+    res: Response,
+    coordinator_privkey: any,
+    backend_pubkey: any
 ) {
     const case_id = req.params.case_uuid;
     // Tell Uploader to Delete File
     let msg: string = req.params.case_uuid;
-    let message: SignedMessage = signMessage(msg);
+    const signature = req.headers.signature as string;
+    let verify = createVerify("sha512");
+    verify.write(msg);
+    verify.end();
+    const isVerified = verify.verify(backend_pubkey, signature, "hex");
+    if (!isVerified) {
+        res.sendStatus(401);
+        return;
+    }
+    // Resign message and send to uploader
+    let sign = createSign("sha512");
+    sign.write(msg);
+    sign.end();
+    const new_signature = sign.sign(coordinator_privkey, "hex");
+    let uploader_url = '';
     fetch(
-        `${UPLOAD_SERVICE}/support/delete_all_case_files/${message.message}`,
+        `${uploader_url}/support/delete_all_case_files/${msg}`,
         {
             headers: new Headers({
-                signature: message.signature,
+                signature: new_signature,
             }),
             method: "DELETE",
         }
     )
-        .then((resp) => {
-            if (resp.ok) res.sendStatus(200);
+        .then(async (resp) => {
+            if (resp.ok) {
+                res.sendStatus(200);
+
+                await execute_sql(
+                    `DELETE FROM files WHERE case_guid = '${case_id}'`
+                );
+                await execute_sql(
+                    `DELETE FROM cases WHERE guid = '${case_id}'`
+                );
+            }
         })
         .catch((err) => {
             console.error(err);
@@ -33,7 +55,4 @@ export async function delete_support_delete_all_cases_files_case_uuid(
         .finally(() => {
             if (!res.closed) res.sendStatus(500);
         });
-
-    await execute_sql(`DELETE FROM files WHERE case_guid = '${case_id}'`);
-    await execute_sql(`DELETE FROM cases WHERE guid = '${case_id}'`);
 }
